Move lucene logic into builder
